New qpsnr.rb wrapper script.
This new script works much like the qpsnr binary but instead of outputing SSIM
and PSNR values to STDOUT it generates a nice HTML page with a flot graph and
links to the individual frames of each SSIM/PSNR value.

This is good to perform a one by one frame comparison of the encoded files vs
the original.